Hoist search query sanitisation into the global namespace
